+++ src/Components/AlterOperation.php
@@ -236,13 +236,23 @@ class AlterOperation extends Component
                     // We have reached the end of ALTER operation and suddenly found
                     // a start to new statement, but have not find a delimiter between them
 
-                    if (!($token->value == 'SET' && $list->tokens[$list->idx - 1]->value == 'CHARACTER')) {
+                    if (!($token->value === 'SET' && $list->tokens[$list->idx - 1]->value === 'CHARACTER')) {
                         $parser->error(
                             'A new statement was found, but no delimiter between it and the previous one.',
                             $token
                         );
                         break;
                     }
+                } elseif (array_key_exists($token->value, self::$DB_OPTIONS)
+                    || (array_key_exists($token->value, self::$TABLE_OPTIONS)
+                    && !self::checkIfColumnDefinitionKeyword($token->value))
+                ) {
+                    // This alter operation has finished, which means a comma was missing before start of new alter operation
+                    $parser->error(
+                        'Missing comma before start of a new alter operation.',
+                        $token
+                    );
+                    break;
                 }
                 $ret->unknown[] = $token;
             }
@@ -275,5 +285,18 @@ class AlterOperation extends Component
         $ret .= TokensList::build($component->unknown);
 
         return $ret;
+    }
+
+    /**
+     * Check if token's value is one of the common keywords
+     * between column and table alteration
+     *
+     * @param string $tokenValue Value of current token
+     */
+    private static function checkIfColumnDefinitionKeyword($tokenValue) {
+        $common_options = array('AUTO_INCREMENT', 'COMMENT', 'DEFAULT');
+        // Since AUTO_INCREMENT and COMMENT can be used for
+        // both table as well as a specific column in the table
+        return in_array($tokenValue, $common_options);
     }
 }